common utils into scraper_utils
